requirements.txt
# Vesuvius Faz 1 - Requirements

# PyTorch (CUDA 11.8 version - adjust based on your CUDA version)
torch>=2.0.0
torchvision>=0.15.0
# torchaudio>=2.0.0  # Optional, not needed for this project

# Core scientific computing
numpy>=1.21.0
scipy>=1.7.0

# Image processing
tifffile>=2021.8.8  # For reading TIFF volumes
scikit-image>=0.19.0  # Optional, for additional image processing
Pillow>=9.0.0  # Image operations

# Configuration and utilities
pyyaml>=6.0  # For YAML config files
tqdm>=4.62.0  # Progress bars

# Visualization (optional but useful)
matplotlib>=3.5.0
tensorboard>=2.10.0  # For training visualization

# Data handling
pandas>=1.3.0  # For data analysis and CSV handling
h5py>=3.7.0  # Optional, for HDF5 file support

# Additional utilities
opencv-python>=4.5.0  # Optional, for advanced image operations

# Kaggle specific (if running on Kaggle)
# kaggle>=1.5.12  # Already installed on Kaggle

# Development tools (optional)
# jupyter>=1.0.0
# ipython>=8.0.0
# black>=22.0.0  # Code formatting
# pytest>=7.0.0  # Testing